requirements.txt
djangorestframework
djangorestframework_simplejwt
psycopg2